rename ObjectToolbar to ActiveFileViewToolbar
Maybe a little clearer that this will show when there's an active file view, even if there's no ActiveObjectId.
